Renamed file to BAFunnel
