+++ dashboard/src/components/funnels/BAFunnel.tsx
